# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.8
  title: River_Applet_Server
  description: |
    Appletサーバ。
host: st-app.riiiver.com
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  JWTToken:
    description: |
      Appから送られるJWTトークン
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /push/notificationToken:
    x-swagger-router-controller: push
    post:
      description: プッシュ通知のNotification Tokenの登録を行う.
      operationId: registerNotificationToken
      security:
        - JWTToken: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - notificationToken
              - appId
              - osType
            properties:
              notificationToken:
                type: string
              appId:
                type: string
              osType:
                $ref: '#/definitions/osType'
      responses:
        '201':
          description: 登録成功
        '400':
          description: パラメータ不正
          schema:
            $ref: '#/definitions/BadRequest'
        '401':
          description: API認証失敗
          schema:
            $ref: '#/definitions/Unauthorized'
  /push/acceptance:
    x-swagger-router-controller: push
    post:
      description: |
        連携先ユーザのデバイスにpush通知を送る
        一回連携できるユーザは１人のみ
      operationId: sendAcceptancePush
      security:
        - JWTToken: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - type
              - ownerUserId
              - userId
              - appId
              - appletId
              - user_preference_id
            properties:
              type:
                type: string
                enum:
                  - link
                  - unlink
                description: link, unlinkのいずれかを指定する。必須項目。
              ownerUserId:
                type: string
                description: 連携元ユーザID
              userId:
                type: string
                description: 連携先ユーザID
              appId:
                type: string
              appletId:
                type: string
              blockId:
                type: string
              user_preference_id:
                type: string
                description: PreferenceID
              message:
                type: string
                description: ★ここのmessageフィールドを追加する
      responses:
        '201':
          description: 送信要求成功
        '400':
          description: パラメータ不正
          schema:
            $ref: '#/definitions/BadRequest'
        '401':
          description: API認証失敗
          schema:
            $ref: '#/definitions/Unauthorized'
        '404':
          description: 存在しないApplet
          schema:
            $ref: '#/definitions/NotFound'
  /push/link:
    x-swagger-router-controller: push
    post:
      description: piece 実行結果を連携先ユーザのデバイスにpush通知を送る
      operationId: sendLinkPush
      security:
        - JWTToken: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - appletId
              - userPreferenceId
              - nextBlockId
              - nextInput
            properties:
              appletId:
                type: string
              userPreferenceId:
                type: string
              nextBlockId:
                type: string
                description: 次に実行されるblockId
              nextInput:
                type: object
                description: 前のblock実行結果（次のblockのinputになる）
      responses:
        '201':
          description: 送信要求成功
        '400':
          description: パラメータ不正
          schema:
            $ref: '#/definitions/BadRequest'
        '401':
          description: API認証失敗
          schema:
            $ref: '#/definitions/Unauthorized'
  /push/privateKey:
    x-swagger-router-controller: push
    post:
      description: FCM 用の秘密鍵を登録・更新する
      operationId: postPrivateKey
      security:
        - JWTToken: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - appId
              - vendorId
              - firebaseProjectId
              - firebasePrivateKey
            properties:
              appId:
                type: string
              vendorId:
                $ref: '#/definitions/VendorId'
              firebaseProjectId:
                type: string
              firebasePrivateKey:
                type: object
                description: Firebase コンソールから取得した秘密鍵
      responses:
        '201':
          description: 登録成功
          schema:
            type: string
        '400':
          description: パラメータ不正
          schema:
            $ref: '#/definitions/BadRequest'
        '401':
          description: API認証失敗
          schema:
            $ref: '#/definitions/Unauthorized'
        '404':
          description: 存在しないApplet
          schema:
            $ref: '#/definitions/NotFound'
...
definitions:
...
  VendorId:
    description: |
      VendorのIDとして用いられる。
    type: string
    pattern: '^[A-Za-z0-9][A-Za-z0-9_]{2,63}$'
    example: yahoo
...
  BadRequest:
    description: |
      400 Bad Requestとともに返却されるエラーオブジェクト。
      API仕様に定めたJSON schemaにマッチしなかった場合などにこのエラーとなる。
      エラー内容は現状では正規化されてはいないが、正規化する場合はDodaiのエラーレスポンス形式に合わせる。
  Unauthorized:
    description: |
      401 Unauthorizedとともに返却されるエラーオブジェクト。
      その他、Dodaiからエラーレスポンスが帰ってきた場合も該当する。
      エラー内容は現状では正規化されてはいないが、正規化する場合はDodaiのエラーレスポンス形式に合わせる。
  NotFound:
    description: |
      404 Not Foundとともに返却されるエラーオブジェクト。
      指定されたリソースが存在しなかった場合にこのエラーとなる。
...
  osType:
    description: blockが動作する対象スマホOS
    type: string
    enum:
      - iOS
      - Android
      - none
...